fix tasks to new structure
