+++ src/org/openstreetmap/josm/plugins/tofix/TofixTask.java
@@ -1,5 +1,6 @@
 package org.openstreetmap.josm.plugins.tofix;
 
+import java.util.List;
 import javax.swing.JOptionPane;
 import org.openstreetmap.josm.Main;
 import org.openstreetmap.josm.actions.downloadtasks.DownloadOsmTask;
@@ -9,6 +10,9 @@ import org.openstreetmap.josm.data.osm.Node;
 import org.openstreetmap.josm.plugins.tofix.bean.AccessToTask;
 import org.openstreetmap.josm.plugins.tofix.bean.items.Item;
 import org.openstreetmap.josm.plugins.tofix.bean.items.ItemKeeprightBean;
+import org.openstreetmap.josm.plugins.tofix.bean.items.ItemKrakatoaBean;
+import org.openstreetmap.josm.plugins.tofix.bean.items.ItemNycbuildingsBean;
+import org.openstreetmap.josm.plugins.tofix.bean.items.ItemTigerdeltaBean;
 import org.openstreetmap.josm.plugins.tofix.bean.items.ItemUnconnectedBean;
 import org.openstreetmap.josm.plugins.tofix.controller.ItemController;
 import org.openstreetmap.josm.plugins.tofix.layer.TofixLayer;
@@ -23,7 +27,6 @@ import static org.openstreetmap.josm.tools.I18n.tr;
 public class TofixTask {
 
     ItemController itemController = new ItemController();
-
     Bounds bounds = null;
     DownloadOsmTask downloadOsmTask = new DownloadOsmTask();
     Node node = null;
@@ -38,13 +41,13 @@ public class TofixTask {
             accessToTask = work_keepright(item.getItemKeeprightBean(), accessToTask);
         }
         if (accessToTask.getTask_source().equals("tigerdelta")) {
-
+            accessToTask = work_tigerdelta(item.getItemTigerdeltaBean(), accessToTask);
         }
         if (accessToTask.getTask_source().equals("nycbuildings")) {
-
+            accessToTask = work_nycbuildings(item.getItemNycbuildingsBean(), accessToTask);
         }
         if (accessToTask.getTask_source().equals("krakatoa")) {
-
+            accessToTask = work_krakatoa(item.getItemKrakatoaBean(), accessToTask);
         }
         return accessToTask;
     }
@@ -68,12 +71,38 @@ public class TofixTask {
 
     }
 
+    private AccessToTask work_nycbuildings(ItemNycbuildingsBean itemNycbuildingsBean, AccessToTask accessToTask) {
+        accessToTask.setKey(itemNycbuildingsBean.getKey());
+        node = itemNycbuildingsBean.getValue().get_node();
+        bounds = new Bounds(node.getCoor().toBBox(Config.bounds).toRectangle());
+        TofixDraw.draw_Node(tofixLayer, node.getCoor());
+        Download.Download(downloadOsmTask, bounds, itemNycbuildingsBean.getValue().osm_obj_id());
+        return accessToTask;
+    }
+
+    private AccessToTask work_tigerdelta(ItemTigerdeltaBean itemTigerdeltaBean, AccessToTask accessToTask) {
+        accessToTask.setKey(itemTigerdeltaBean.getKey());
+        List<List<Node>> list = itemTigerdeltaBean.getValue().get_nodes();
+        node = new Node(new LatLon(list.get(0).get(0).getCoor().lat(), list.get(0).get(0).getCoor().lon()));
+        bounds = new Bounds(node.getCoor().toBBox(Config.bounds).toRectangle());
+        TofixDraw.draw_line(tofixLayer, node.getCoor(), list);
+        Download.Download(downloadOsmTask, bounds, 0x0L);//0x0L = null porque no exixte el id del objeto
+        return accessToTask;
+    }
+
+    private AccessToTask work_krakatoa(ItemKrakatoaBean itemkrakatoaBean, AccessToTask accessToTask) {
+        accessToTask.setKey(itemkrakatoaBean.getKey());
+        List<Node> list = itemkrakatoaBean.getValue().get_nodes();
+        node = new Node(new LatLon(list.get(0).getCoor().lat(), list.get(0).getCoor().lon()));
+        bounds = new Bounds(node.getCoor().toBBox(Config.bounds).toRectangle());
+        TofixDraw.draw_nodes(tofixLayer, node.getCoor(), list);
+        Download.Download(downloadOsmTask, bounds, 0x0L);//0x0L = null porque no exixte el id del objeto
+        return accessToTask;
+    }
+
     public void task_complete(Item item, AccessToTask accessToTask) {
         String message = "Task : " + accessToTask.getTask_name() + " was completed\n"
                 + "Total items : " + item.getTaskCompleteBean().getMessage().getValue().getTotal();
-        //  + "Fixed : " + item.getTaskCompleteBean().getMessage().getValue().getFix() + "\n"
-        //  + "Not Error : " + item.getTaskCompleteBean().getMessage().getValue().getNoterror() + "\n"
-        //  + "Skip : " + item.getTaskCompleteBean().getMessage().getValue().getSkip() + "\n";
         JOptionPane.showMessageDialog(Main.panel, tr(message));
     }
