+++ src/org/openstreetmap/josm/plugins/tofix/bean/items/ItemNycbuildingsBean.java
@@ -1,17 +1,16 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package org.openstreetmap.josm.plugins.tofix.bean.items;
+
+import org.openstreetmap.josm.data.coor.LatLon;
+import org.openstreetmap.josm.data.osm.Node;
 
 /**
  *
  * @author ruben
  */
 public class ItemNycbuildingsBean {
-   private  String key;
-   private  ItemNycbuildingsValueBean value;
+
+    private String key;
+    private Value value;
 
     public String getKey() {
         return key;
@@ -21,13 +20,54 @@ public class ItemNycbuildingsBean {
         this.key = key;
     }
 
-    public ItemNycbuildingsValueBean getValue() {
+    public Value getValue() {
         return value;
     }
 
-    public void setValue(ItemNycbuildingsValueBean value) {
+    public void setValue(Value value) {
         this.value = value;
     }
-   
-}
 
+    public class Value {
+
+        private Double lat;
+        private Double lon;
+        private String elems;
+
+        public Double getLat() {
+            return lat;
+        }
+
+        public void setLat(Double lat) {
+            this.lat = lat;
+        }
+
+        public Double getLon() {
+            return lon;
+        }
+
+        public void setLon(Double lon) {
+            this.lon = lon;
+        }
+
+        public String getElems() {
+            return elems;
+        }
+
+        public void setElems(String elems) {
+            this.elems = elems;
+        }
+
+        public Long osm_obj_id() {
+            String[] arr = getElems().replace("way", "").split("_");
+            return Long.valueOf(arr[0]);
+        }
+        public Node get_node(){
+            Node node = new Node(new LatLon(getLat(),getLon()));
+            return  node;
+            
+        }        
+
+    }
+
+}
