--- src/org/openstreetmap/josm/plugins/tofix/bean/items/ItemNycbuildingsValueBean.java
@@ -1,39 +1,0 @@
-package org.openstreetmap.josm.plugins.tofix.bean.items;
-
-/**
- *
- * @author ruben
- */
-public class ItemNycbuildingsValueBean {
-
-    private Double lat;
-    private Double lon;
-    private String elems;
-
-    public Double getLat() {
-        return lat;
-    }
-
-    public void setLat(Double lat) {
-        this.lat = lat;
-    }
-
-    public Double getLon() {
-        return lon;
-    }
-
-    public void setLon(Double lon) {
-        this.lon = lon;
-    }
-
-    public String getElems() {
-        return elems;
-    }
-
-    public void setElems(String elems) {
-        this.elems = elems;
-    }
-
-
-
-}
